+++ packages/code-du-travail-frontend/src/outils/IndemniteLicenciement/steps/Anciennete/store/validator/absenceProlonge.ts
@@ -1,0 +1,15 @@
+import { AncienneteStoreError, AncienneteStoreInput } from "../types";
+
+export const getAbsenceProlongeErrors = (
+  state: AncienneteStoreInput
+): Partial<AncienneteStoreError> => {
+  let errors: AncienneteStoreError = {};
+
+  if (state.hasAbsenceProlonge === undefined) {
+    errors.errorAbsenceProlonge = "Vous devez répondre à cette question";
+  } else {
+    errors.errorAbsenceProlonge = undefined;
+  }
+
+  return errors;
+};
