+++ packages/code-du-travail-frontend/src/outils/IndemniteLicenciement/steps/Anciennete/store/validator/dateEntree.ts
@@ -1,0 +1,19 @@
+import { isValidDate } from "../../../../../../lib";
+import { AncienneteStoreError, AncienneteStoreInput } from "../types";
+
+export const getDateEntreeErrors = (
+  state: AncienneteStoreInput
+): Partial<AncienneteStoreError> => {
+  let errors: AncienneteStoreError = {};
+
+  // Date d'entrée
+  if (!state.dateEntree) {
+    errors.errorDateEntree = "Veuillez saisir cette date";
+  } else if (!isValidDate(state.dateEntree)) {
+    errors.errorDateEntree = "La date d'entrée est invalide";
+  } else {
+    errors.errorDateEntree = undefined;
+  }
+
+  return errors;
+};
